Chore: nettoyer les imports du module d'export
